requirements.txt
bs4==0.0.1
google-api-python-client==2.79.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
lxml==4.9.2
python-dotenv==1.0.0
requests==2.28.2
selenium==4.8.2
undetected-chromedriver==3.4.6